:sparkles: 대시보드 API composite aggregation after 검색 적용

See merge request s10-final/S10P31S106!33

+++ Backend/src/main/java/com/dfg/semento/dto/response/StateHourlyResponse.java
@@ -14,5 +14,5 @@ public class StateHourlyResponse {
 
 	private int hour;
 
-	private long count;
+	private double count;
 }
